% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlsb.R
\name{read_xlsb}
\alias{read_xlsb}
\title{Loads a XLSB file}
\usage{
read_xlsb(x)
}
\arguments{
\item{x}{A filename or filepath.}
}
\value{
A dataframe containing data stored in the file.
}
\description{
This is the basic function of the package, wich reads the xlsb files as a database using RODBC.
}
\examples{
DT <- read_xlsb("data/march.xlsb")
DT <- read_xlsb("march.xlsb")
}
